requirements.txt
flask==3.0.0
flask-sqlalchemy==3.1.1
marshmallow<4.0.0
flask-marshmallow==0.15.0
flask-cors==4.0.0
marshmallow-sqlalchemy==0.30.0
pymysql==1.1.0
python-dotenv==1.0.0
pytest==7.4.4
pytest-flask==1.3.0
